==> notebooks/1_data_formatting.ipynb <==
## IPL data analysis

- Format the deliveries file and matches file

- Merge deliveries and matches file

# Import libraries

import pandas as pd

# Libraries version

print(f'Pandas Version - {pd.__version__}')

# Pandas display options

pd.options.display.float_format = '{:.2f}'.format
pd.set_option('display.max_columns', None)
pd.set_option('display.max_colwidth', None)
pd.set_option('display.max_rows', None)

# Read file

file_path = 'https://raw.githubusercontent.com/nagaraj-datascientist/ipl_data_analysis/master/data/'

deliveries_df = pd.read_csv(f'{file_path}deliveries.csv')
matches_df = pd.read_csv(f'{file_path}matches.csv')

# Glimpse of deliveries data

deliveries_df.head()

# Glimpse of matches data

matches_df.head()

# Info of deliveries

deliveries_df.info(memory_usage='deep')

# Info of matches

matches_df.info(memory_usage='deep')

### Format Matches file

Memory usage of matches file is 676 kb.

Assign alias name to the ipl teams and change the datatype.


#### Team name alias

- Assign same name if one team shares 2 names

# Unique team name

matches_df['team1'].unique()

# Replace team name

matches_df.replace('Rising Pune Supergiant', 'Rising Pune Supergiants', inplace=True)
matches_df.replace('Delhi Daredevils', 'Delhi Capitals', inplace=True)
matches_df.replace('Deccan Chargers', 'Sunrisers Hyderabad', inplace=True)

# Alias name function

def team_alias(team):
    if team=='Sunrisers Hyderabad':
        return 'SRH'
    elif team=='Mumbai Indians':
        return 'MI'
    elif team=='Gujarat Lions':
        return 'GL'
    elif team=='Rising Pune Supergiants'or team=='Pune Warriors':
        return 'RPSG'
    elif team=='Royal Challengers Bangalore':
        return 'RCB'
    elif team=='Kolkata Knight Riders':
        return 'KKR'
    elif team=='Delhi Capitals':
        return 'DD'
    elif team=='Kings XI Punjab':
        return 'KXIP'
    elif team=='Chennai Super Kings':
        return 'CSK'
    elif team=='Rajasthan Royals':
        return 'RR'
    elif team=='Kochi Tuskers Kerala':
        return 'KTK'
    else:
        return team

# Alias name assignment for matches data

teams_columns = ['team1', 'team2', 'toss_winner', 'winner']

for t in teams_columns:
    matches_df[t] = matches_df[t].apply(team_alias)


# Unique team names

matches_df['team1'].unique()

#### Datatype change

- category - city, team1, team2, toss_winner, toss_decision, result, winner

- datetime - date


# Data type changes

matches_df = matches_df.astype(
    {'city': 'category', 'team1': 'category', 'team2': 'category',
     'toss_winner': 'category', 'toss_decision': 'category',
     'result': 'category', 'winner': 'category', 'date': 'datetime64[ns]'}) 

# Info of matches

matches_df.info(memory_usage='deep')

Now matches if fomratted and shaped, memory usage is ~285 kb.

### Format Deliveries file

Memory usage of deliveries file is 94.5 MB.

Assign alias name to the ipl teams and change the datatype.

# Replace team name

deliveries_df.replace('Rising Pune Supergiant', 'Rising Pune Supergiants', inplace=True)
deliveries_df.replace('Delhi Daredevils', 'Delhi Capitals', inplace=True)
deliveries_df.replace('Deccan Chargers', 'Sunrisers Hyderabad', inplace=True)

# Alias name assignment for matches data

teams_columns = ['batting_team', 'bowling_team']

for t in teams_columns:
    deliveries_df[t] = deliveries_df[t].apply(team_alias)


### Datatype change

- datatype changes - batting_team, bowling_team


# datatype change

deliveries_df = deliveries_df.astype({'batting_team': 'category',
                                      'bowling_team': 'category'})

# Info of deliveries

deliveries_df.info(memory_usage='deep')

Now memory usage of deliveries file is ~ 69 MB

### Venue name

Venue shares more than 1 name in matches file

# Number of unique venue

matches_df['venue'].nunique()

sorted(matches_df['venue'].unique())

# Replacing venue names

matches_df['venue']=matches_df['venue'].str.replace('Feroz Shah Kotla Ground','Feroz Shah Kotla')
matches_df['venue']=matches_df['venue'].str.replace('M. Chinnaswamy Stadium', 'M Chinnaswamy Stadium')
matches_df['venue']=matches_df['venue'].str.replace('M. A. Chidambaram Stadium', 'MA Chidambaram Stadium, Chepauk')
matches_df['venue']=matches_df['venue'].str.replace('Rajiv Gandhi Intl. Cricket Stadium', 
                                                    'Rajiv Gandhi International Stadium, Uppal')
matches_df['venue']=matches_df['venue'].str.replace('Punjab Cricket Association IS Bindra Stadium, Mohali', 
                                'Punjab Cricket Association Stadium, Mohali')

matches_df['venue']=matches_df['venue'].str.replace('IS Bindra Stadium', 
                                'Punjab Cricket Association Stadium, Mohali')


matches_df['venue'].nunique()

### Merge the 2 files

- Merge matches and deliveries data into one file for further analysis.

# Merge 2 files

ipl_df = pd.merge(deliveries_df, matches_df, left_on='match_id', right_on='id')

# Save merged content into file

ipl_df.to_csv('deliveries_and_matches.csv', index=False)

# Save formatted file

deliveries_df.to_csv('deliveries_formatted.csv', index=False)
matches_df.to_csv('matches_formatted.csv', index=False)
