==> notebooks/2_all_time_records.ipynb <==
# IPL data analysis

- Replicate same reports as [iplt20.com](https://www.iplt20.com/stats/all-time).


# Import the libraries

import pandas as pd

# Libraries version

print(f'Pandas Version - {pd.__version__}')

# Pandas display options

pd.options.display.float_format = '{:.2f}'.format
pd.set_option('display.max_columns', None)
pd.set_option('display.max_colwidth', None)
pd.set_option('display.max_rows', None)

# Read file

file_path = 'https://raw.githubusercontent.com/nagaraj-datascientist/ipl_data_analysis/master/data/'

ipl_df = pd.read_csv(f'{file_path}deliveries_and_matches.csv', low_memory=False)


# Glimpse of data

ipl_df.head()

# Info of ipl_df

ipl_df.info(memory_usage='deep')

# Data type changes

ipl_df = ipl_df.astype({'city': 'category', 'team1': 'category',
                        'team2': 'category', 'toss_winner': 'category',
                        'toss_decision': 'category', 'result': 'category',
                        'winner': 'category', 'date': 'datetime64[ns]',
                        'batting_team': 'category', 'bowling_team': 'category'}) 

# Info of ipl_df

ipl_df.info(memory_usage='deep')

## All time Batting leaders

Most Runs, Most Sixes, Highest Score, Best Strike Rate, Most Fours. 

*Refer iplt20 [here](https://www.iplt20.com/stats/all-time/most-runs)*

# Most Runs

most_runs_df = ipl_df.groupby(by='batsman').agg(Inns = ('match_id', 'nunique'),
                                 Runs = ('batsman_runs','sum'),
                                 BF = ('batsman_runs', 'count')).reset_index().sort_values(by='Runs', ascending=False)

most_runs_df.head()

# Most Sixes

most_sixes_df = ipl_df[ipl_df['batsman_runs']>=6].groupby(by='batsman').agg(Sixes = ('batsman_runs', 'count')).reset_index().sort_values(by='Sixes', ascending=False)

most_sixes_df.head()

# Highest scores innings

high_score_inn_df = ipl_df.groupby(by=['batsman', 'match_id']).agg(HS = ('batsman_runs', 'sum')).reset_index().sort_values(by='HS', ascending=False)

high_score_inn_df.head()

# Most Fours

most_fours_df = ipl_df[ipl_df['batsman_runs'] == 4].groupby(by='batsman').agg(Fours = ('batsman_runs', 'count')).reset_index().sort_values(by='Fours', ascending=False)

most_fours_df.head()

# Batsman Strike Rate

bat_strike_rate_df = ipl_df.groupby(by='batsman').agg(Runs = ('batsman_runs','sum'),
                                 BF = ('batsman_runs', 'count')).reset_index()

bat_strike_rate_df['SR'] = (bat_strike_rate_df['Runs'] / bat_strike_rate_df['BF'])*100

bat_strike_rate_df[bat_strike_rate_df['Runs']> 250].sort_values(by='SR', ascending=False).head()


## All time Bowling leaders

Most Wickets, Best Bowling Innings, Best Bowling Average, Best Bowling Economy, Most Dot Balls 

# Most Wickets

bowler_wicket_scenario = ['caught', 'bowled', 'lbw', 'stumped',
                          'caught and bowled', 'hit wicket']

most_wicket_df = ipl_df[ipl_df['dismissal_kind'].isin(bowler_wicket_scenario)].groupby(by='bowler').agg(Wkts=('dismissal_kind', 'count')).reset_index().sort_values('Wkts', ascending=False)

most_wicket_df.head()

# Best Bowling Innings

bowler_wkt_inn_df = ipl_df[ipl_df['dismissal_kind'].isin(bowler_wicket_scenario)].groupby(by=['bowler', 'match_id']).agg(Wkts = ('dismissal_kind', 'count')).reset_index()

ipl_df['bowler_runs']= ipl_df['extra_runs'] + ipl_df['batsman_runs']

bowler_run_inn_df = ipl_df.groupby(by=['match_id', 'bowler']).agg(Runs=('bowler_runs', 'sum')).reset_index()

best_bowling_inn_df = pd.merge(bowler_wkt_inn_df, bowler_run_inn_df, on=['match_id', 'bowler'])

best_bowling_inn_df.sort_values(['Wkts', 'Runs'], ascending=[False, True]).head()

# Most Dot Balls

most_dot_balls_df = ipl_df[ipl_df['total_runs'] == 0].groupby('bowler').agg(Dots=('ball', 'count')).reset_index().sort_values('Dots', ascending=False)
most_dot_balls_df.head()
