==> notebooks/3_other_analysis.ipynb <==
# IPL data analysis

Slice and dice the data section-wise

Sections:

- Venue

- Team, etc

# Import the libraries

import pandas as pd

# Libraries version

print(f'Pandas Version - {pd.__version__}')

# Pandas display options

pd.options.display.float_format = '{:.2f}'.format
pd.set_option('display.max_columns', None)
pd.set_option('display.max_colwidth', None)
pd.set_option('display.max_rows', None)

# Read file

file_path = 'https://raw.githubusercontent.com/nagaraj-datascientist/ipl_data_analysis/master/data/'

ipl_df = pd.read_csv(f'{file_path}deliveries_and_matches.csv', low_memory=False)

matches_df = pd.read_csv(f'{file_path}matches_formatted.csv', low_memory=False)


# Glimpse of data

ipl_df.head()

# Match count overall

ipl_df['match_id'].nunique()

# Results

ipl_df.groupby('result').agg(Results = ('match_id', 'nunique')).reset_index().sort_values('Results', ascending=False)

# Match counts by season

ipl_df.groupby('season').agg(match_count = ('match_id', 'nunique')).reset_index().sort_values('season', ascending=False)

# Team head to head

teams = ['CSK', 'KXIP']

team_head_to_head = ipl_df[(((ipl_df.team1==teams[0]) & (ipl_df.team2==teams[1])) | ((ipl_df.team1==teams[1]) & (ipl_df.team2==teams[0])))]

team_head_to_head.groupby(['winner', 'result']).agg(result_count = ('match_id', 'nunique')).reset_index()

# Unique innings

ipl_df.inning.unique()

# More than 2 innings

ipl_df[ipl_df['inning'] >2 ].head()

# History board

ipl_history_df = ipl_df[ipl_df.inning.isin([1,2])].groupby(['match_id', 'season','inning','batting_team']).agg(score=('total_runs','sum'), wickets =('player_dismissed','count')).reset_index()

ipl_history_df = ipl_history_df.pivot_table(index=['match_id', 'season'], 
                           columns='inning',
                           values=['batting_team', 'score','wickets'],
                           aggfunc='first').reset_index()

ipl_history_df.columns = [f'{i}_{j}' if j != '' else f'{i}' for i,j in ipl_history_df.columns]

ipl_history_df.head()

# Info of ipl_history_df

ipl_history_df.info()

# Formatting ipl_history_df

ipl_history_df.fillna({'batting_team_2 ': 'Not Available',
                       'score_2': 0,
                       'wickets_2': 1}, inplace=True)

ipl_history_df = ipl_history_df.astype({'score_1': 'int64', 'score_2': 'int64', 
                                        'wickets_1': 'int64', 'wickets_2': 'int64'})

ipl_history_df.info()

# Team winning rate

team1 = matches_df.groupby('team1')['team1'].count()
team2 = matches_df.groupby('team2')['team2'].count()

team_matches = pd.DataFrame(team1 + team2).reset_index()

team_wins = matches_df[matches_df.result == 'normal'].groupby('winner').agg(no_of_won=('winner','count')).reset_index()

team_matches_wins = pd.merge(team_matches, team_wins, left_on='team1', right_on='winner')

team_matches_wins = team_matches_wins.rename(columns={'team1': 'Team', 0: 'total_matches'}).drop(columns='winner')

team_matches_wins['win_rate'] = team_matches_wins['no_of_won'] / team_matches_wins['total_matches']

team_matches_wins.sort_values('win_rate', ascending=False).head()

# Dockout counts

batsman_runs_matches = ipl_df.groupby(by=['match_id', 'batsman']).agg(batsman_runs_match = ('batsman_runs', 'sum')).reset_index()

batsman_runs_matches[batsman_runs_matches['batsman_runs_match']==0].groupby('batsman').agg(duck_out = ('batsman', 'count')).reset_index().sort_values('duck_out', ascending=False).head()

# Hit wicket

ipl_df[ipl_df.dismissal_kind == 'hit wicket'].groupby(['batsman']).agg('count').reset_index()[['batsman', 'player_dismissed']]


# Obstructing the field

ipl_df[ipl_df.dismissal_kind == 'obstructing the field'].groupby(['batsman']).agg('count').reset_index()[['batsman', 'player_dismissed']]


# Retired hurt

ipl_df[ipl_df.dismissal_kind == 'retired hurt'].groupby(['batsman']).agg('count').reset_index()[['batsman', 'player_dismissed']]


# Man of the match

matches_df.groupby(['player_of_match']).agg(man_of_match = ('player_of_match','count')).reset_index().sort_values(by='man_of_match', ascending=False).head()


# No of catches 

ipl_df[ipl_df.dismissal_kind == 'caught'].groupby(['fielder']).agg(no_of_catches = ('fielder', 'count')).reset_index().sort_values(by='no_of_catches', ascending=False).head()


# Run out count

ipl_df[ipl_df.dismissal_kind == 'run out'].groupby(['fielder']).agg(run_out_count = ('fielder', 'count')).reset_index().sort_values(by='run_out_count', ascending=False).head()


# Stumped

ipl_df[ipl_df.dismissal_kind == 'stumped'].groupby(['fielder']).agg(no_of_stumped = ('fielder', 'count')).reset_index().sort_values(by='no_of_stumped', ascending=False).head()


# Average score first innings and venue

first_innings = ipl_df[ipl_df.inning == 1]
second_innings = ipl_df[ipl_df.inning == 2]

first_ings_venue = first_innings.groupby(['venue','match_id']).agg(total_score_per_innings = ('total_runs', 'sum')).reset_index()

first_ings_venue.groupby('venue').agg(avg_first_innings_score_venue = ('total_score_per_innings', 'mean')).sort_values(by ='avg_first_innings_score_venue',ascending=False).reset_index().head()

# Average score second innings and venue

second_ings_venue = second_innings.groupby(['venue','match_id']).agg(total_score_per_innings = ('total_runs', 'sum')).reset_index()

second_ings_venue.groupby('venue').agg(avg_second_innings_score_venue = ('total_score_per_innings', 'mean')).sort_values(by ='avg_second_innings_score_venue',ascending=False).reset_index().head()
